requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.41
python-dotenv
fastapi 
google-auth-oauthlib 
google-api-python-client 
caldav
